Abstracted models into individual files
