+++ Concertino/Models/AddWork.swift
@@ -1,0 +1,12 @@
+//
+//  AddWork.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct AddWork: Codable {
+    var list: [String]
+    var composerworks: [String]
+}
