+++ Concertino/Models/Composer.swift
@@ -1,0 +1,19 @@
+//
+//  Composer.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+import SwiftUI
+
+struct Composer: Codable, Identifiable {
+    let id: String
+    var name: String
+    var complete_name: String
+    var birth: String?
+    var death: String?
+    var epoch: String
+    var portrait: URL?
+}
