+++ Concertino/Models/Composers.swift
@@ -1,0 +1,11 @@
+//
+//  Composers.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct Composers: Codable {
+    var composers: [Composer]?
+}
