+++ Concertino/Models/Genres.swift
@@ -1,0 +1,11 @@
+//
+//  Genres.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct Genres: Codable {
+    var genres: [String]?
+}
