+++ Concertino/Models/PlaylistRecording.swift
@@ -1,0 +1,11 @@
+//
+//  PlaylistRecording.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct PlaylistRecording: Codable {
+    var list: [Playlist]
+}
