+++ Concertino/Models/PlaylistRecordings.swift
@@ -1,0 +1,11 @@
+//
+//  PlaylistRecordings.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct PlaylistRecordings: Codable {
+    var recordings: [Recording]?
+}
