+++ Concertino/Models/PlaylistSummary.swift
@@ -1,0 +1,12 @@
+//
+//  PlaylistSummary.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct PlaylistSummary: Codable {
+    var composers: PlaylistComposers
+    var works: PlaylistWorks
+}
