+++ Concertino/Models/RadioStationPlaylist.swift
@@ -1,0 +1,15 @@
+//
+//  RadioStationPlaylist.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+import SwiftUI
+
+struct RadioStationPlaylist: Codable {
+    var id: String
+    var name: String
+    var cover: URL
+}
