+++ Concertino/Models/RadioStationPlaylists.swift
@@ -1,0 +1,11 @@
+//
+//  RadioStationPlaylists.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct RadioStationPlaylists: Codable {
+    var playlists: [RadioStationPlaylist]
+}
