+++ Concertino/Models/Recommendation.swift
@@ -1,0 +1,12 @@
+//
+//  Recommendation.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct Recommendation: Codable, Identifiable {
+    var id: String
+    var type: String
+}
