+++ Concertino/Models/Recommendations.swift
@@ -1,0 +1,11 @@
+//
+//  Recommendations.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct Recommendations: Codable {
+    var data: [Recommendation]
+}
