+++ Concertino/Models/Recordings.swift
@@ -1,0 +1,12 @@
+//
+//  Recordings.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct Recordings: Codable {
+    var recordings: [Recording]?
+    var next: String?
+}
