+++ Concertino/Models/ShortRecording.swift
@@ -1,0 +1,14 @@
+//
+//  ShortRecording.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct ShortRecording: Codable {
+    var id: String
+    var work_id: String?
+    var apple_albumid: String?
+    var set: String?
+}
