+++ Concertino/Models/ShortRecordingDetail.swift
@@ -1,0 +1,11 @@
+//
+//  ShortRecordingDetail.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct ShortRecordingDetail: Codable {
+    var recording: ShortRecording
+}
