+++ Concertino/Models/Supporters.swift
@@ -1,0 +1,11 @@
+//
+//  Supporters.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct Supporters: Codable {
+    var patrons: [String]
+}
