+++ Concertino/Models/TrackQueued.swift
@@ -1,0 +1,12 @@
+//
+//  File.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct TrackQueued: Codable {
+    var apple_trackid: String
+    var recording_id: String
+}
