+++ Concertino/Models/Work.swift
@@ -1,0 +1,17 @@
+//
+//  Work.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct Work: Codable, Identifiable {
+    let id: String
+    var title: String
+    var subtitle: String?
+    var genre: String
+    var recommended: String?
+    var popular: String?
+    var composer: Composer?
+}
