+++ Concertino/Models/Works.swift
@@ -1,0 +1,11 @@
+//
+//  Works.swift
+//  Concertino
+//
+//  Created by Kyle Dold on 09/01/2021.
+//  Copyright © 2021 Open Opus. All rights reserved.
+//
+
+struct Works: Codable {
+    var works: [Work]?
+}
